+++ backend/Core/Boundaries/JwtTokenProvider.cs
@@ -1,0 +1,35 @@
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.IdentityModel.Tokens;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
+
+namespace Core.Boundaries;
+
+public class JwtTokenProvider(IConfiguration configuration)
+{
+    private readonly IConfiguration _configuration = configuration;
+
+    public string GenerateToken(Guid userId, string email)
+    {
+        var secretKey = _configuration["Jwt:Secret"];
+        var key = Encoding.ASCII.GetBytes(secretKey);
+
+        var tokenHandler = new JwtSecurityTokenHandler();
+        var tokenDescriptor = new SecurityTokenDescriptor
+        {
+            Subject = new ClaimsIdentity(new[]
+            {
+                new Claim(ClaimTypes.NameIdentifier, userId.ToString()),
+                new Claim(ClaimTypes.Email, email)
+            }),
+            Expires = DateTime.UtcNow.AddHours(2),
+            SigningCredentials = new SigningCredentials(
+                new SymmetricSecurityKey(key),
+                SecurityAlgorithms.HmacSha256Signature)
+        };
+
+        var token = tokenHandler.CreateToken(tokenDescriptor);
+        return tokenHandler.WriteToken(token);
+    }
+}
