--- backend/Infra/Repositories/JwtTokenProvider.cs
@@ -1,66 +1,0 @@
-﻿using Core.Domain;
-using Core.Interfaces;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.Extensions.Configuration;
-using Microsoft.IdentityModel.Tokens;
-using Microsoft.OpenApi.Models;
-using Swashbuckle.AspNetCore.SwaggerGen;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Text;
-
-namespace Infra.Repositories;
-
-public class JwtTokenProvider(
-    IConfiguration configuration
-) : IJwtTokenProvider
-{
-    private readonly IConfiguration _configuration = configuration;
-
-    public void Apply(OpenApiOperation operation, OperationFilterContext context)
-    {
-        var authRequirement = context.ApiDescription.ActionDescriptor.EndpointMetadata
-            .OfType<AuthorizeAttribute>()
-            .Any();
-
-        if (authRequirement)
-        {
-            operation.Security = new List<OpenApiSecurityRequirement>
-        {
-            new OpenApiSecurityRequirement
-            {
-                {
-                    new OpenApiSecurityScheme
-                    {
-                        Reference = new OpenApiReference { Type = ReferenceType.SecurityScheme, Id = "Bearer" }
-                    },
-                    new List<string>()
-                }
-            }
-        };
-        }
-    }
-
-    public string GenerateToken(Guid id, string email)
-    {
-        var secretKey = _configuration["Jwt:Secret"];
-        var key = Encoding.ASCII.GetBytes(secretKey);
-
-        var tokenHandler = new JwtSecurityTokenHandler();
-        var tokenDescriptor = new SecurityTokenDescriptor
-        {
-            Subject = new ClaimsIdentity(new[]
-            {
-                new Claim(ClaimTypes.NameIdentifier, id.ToString()),
-                new Claim(ClaimTypes.Email, email)
-            }),
-            Expires = DateTime.UtcNow.AddHours(2),
-            SigningCredentials = new SigningCredentials(
-                new SymmetricSecurityKey(key),
-                SecurityAlgorithms.HmacSha256Signature)
-        };
-
-        var token = tokenHandler.CreateToken(tokenDescriptor);
-        return tokenHandler.WriteToken(token);
-    }
-}
